requirements.txt
Flask==0.10.1
google-cloud-storage
google-cloud-datastore==1.15.0
datetime
google-auth==1.21.1
requests==2.24.0